Rename repository and adjust template names
As part of adding this repository to the official list of community
devcontainer template repositories, I want to make it more clear that
the initial set of templates that are part of this repository are
_barebones_ to avoid confusion with other templates (in particular the
official ones).

My goals have also evolved beyond just having barebones containers to
include some more opinionated ones in the future (e.g. a template for
how I set up new Rails projects), so "lightweight" in the repo name
doesn't really make sense anymore.

- Update existing templates to be called "barebones-X"
- Adjust tests and workflows as appropriate
- Update repository name as appropriate
- Update README with more suitable descriptions